+++ package.json
@@ -4,6 +4,7 @@
   "description": "ife: Question Management System",
   "main": "index.js",
   "scripts": {
+    "server": "node server.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
