+++ src/page_editor/question/index.js
@@ -1,0 +1,3 @@
+exports.multiplechoice = require('./multiplechoice')
+exports.onechoice      = require('./onechoice')
+exports.text           = require('./text')
