+++ src/page_editor/question/multiplechoice/index.js
@@ -1,0 +1,25 @@
+require('./index.styl')
+
+var template  = require('./template.html')
+var Model     = require('../../model')
+
+var multipleChoice = Backbone.View.extend({
+	tagName: 'h4',
+
+    className: 'question-number',
+
+    events:{
+    },
+
+    initialize: function (options) {
+        this.model = new Model(options)
+    },
+
+    render: function () {
+        this.$el.html(template(this.model.toJSON()))
+
+        return this
+    }
+})
+
+module.exports = multipleChoice
