+++ src/page_editor/question/text/index.js
@@ -1,9 +1,9 @@
 require('./index.styl')
 
 var template  = require('./template.html')
-var Model = require('../model')
+var Model = require('../../model')
 
-var text_question = Backbone.view.extend({
+var text = Backbone.View.extend({
 	tagName: 'h4',
     className: 'question-number',
 
@@ -21,4 +21,4 @@ var text_question = Backbone.view.extend({
   }
 })
 
-module.exports = text_question
+module.exports = text
