--- src/page_editor/question_type/index.js
@@ -1,3 +1,0 @@
-exports.multiplechoice = require('./multiplechoice_question')
-exports.onechoice      = require('./onechoice_question')
-exports.text           = require('./text_question')
