+++ src/page_editor/template.html
@@ -1,4 +1,3 @@
-<div>这是editor</div>
 <div class="box">
 	<h2>这里是标题</h2>
 	<div class="empty"></div>
