+++ src/page_new/template.html
@@ -1,4 +1,3 @@
-<div>这是new</div>
 <div class="newPage">
 	<a href="#add"><button class="add">+新建问卷</button></a>
 </div>
